+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:example/composables/useTheme.dart';
 import 'package:example/middleware/auth_middleware.dart';
 import 'package:flutter/material.dart';
 import 'package:flutterwind_core/flutterwind.dart';
@@ -8,10 +9,13 @@ import 'package:example/generated/routes.dart';
 
 void main() async {
   await VortexRouter.discoverRoutes(
-    projectDirectory: Directory('/Volumes/EVILRATT/Innovative Projects/vortex/example'),
+    projectDirectory: Directory(
+      '/Volumes/EVILRATT/Innovative Projects/vortex/example',
+    ),
   );
 
   initializeRoutes();
+  registerThemeComposable();
   MiddlewareRegistry.register('auth', AuthMiddleware());
   runApp(const MyApp());
 }
@@ -27,8 +31,10 @@ class _MyAppState extends State<MyApp> with CompositionMixin {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    final isDarkMode = useRef<bool>('isDarkMode', false);
-
+    // final isDarkMode = useRef<bool>('isDarkMode', false);
+    final useThemeComposable = Vortex.getComposable('useTheme');
+    final themeState = useThemeComposable(context);
+    final isDarkMode = themeState.isDarkMode;
     // Set up watchers
     watch(isDarkMode, () => isDarkMode.value, (newValue, oldValue) {
       Log.i(
@@ -43,30 +49,35 @@ class _MyAppState extends State<MyApp> with CompositionMixin {
 
     return Vortex(
       child: FlutterWind(
-        child: MaterialApp(
-          title: 'Vortex Demo',
-          theme: ThemeData(
-            colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-            useMaterial3: true,
-            brightness: Brightness.light,
-          ),
-          darkTheme: ThemeData(
-            colorScheme: ColorScheme.fromSeed(
-              seedColor: Colors.deepPurple,
-              brightness: Brightness.dark,
-            ),
-            useMaterial3: true,
-            brightness: Brightness.dark,
-          ),
-          themeMode: isDarkMode.value ? ThemeMode.dark : ThemeMode.light,
-          initialRoute: '/',
-          onGenerateInitialRoutes:
-              (initialRoute) => [
-                VortexRouter.initialRouteHandler(
-                  RouteSettings(name: initialRoute),
+        child: ReactiveBuilder(
+          dependencies: [isDarkMode],
+          builder: (context) {
+            return MaterialApp(
+              title: 'Vortex Demo',
+              theme: ThemeData(
+                colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+                useMaterial3: true,
+                brightness: Brightness.light,
+              ),
+              darkTheme: ThemeData(
+                colorScheme: ColorScheme.fromSeed(
+                  seedColor: Colors.deepPurple,
+                  brightness: Brightness.dark,
                 ),
-              ],
-          onGenerateRoute: VortexRouter.onGenerateRoute,
+                useMaterial3: true,
+                brightness: Brightness.dark,
+              ),
+              themeMode:  isDarkMode.value ? ThemeMode.dark : ThemeMode.light,
+              initialRoute: '/',
+              onGenerateInitialRoutes:
+                  (initialRoute) => [
+                    VortexRouter.initialRouteHandler(
+                      RouteSettings(name: initialRoute),
+                    ),
+                  ],
+              onGenerateRoute: VortexRouter.onGenerateRoute,
+            );
+          }
         ),
       ),
     );