+++ lib/src/composables/index.dart
@@ -1,1 +1,2 @@
 export 'use_fetch.dart';
+export 'registry.dart';